refactor(cli): folder structure
