+++ packages/cli/src/commands/index.ts
@@ -1,0 +1,2 @@
+export * from "./manage";
+export * from "./staticBuild";
